requirements.txt
PilloW ==8.4.0
pandas ==1.1.3